requirements.txt
kafka-python
elasticsearch
pandas
cassandra-driver